+++ code/packages/rust/window-appkit/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this package will be documented in this file.
 
+## Unreleased
+
+### Added
+
+- An AppKit event view that translates wheel input to normalized
+  `WindowEvent::Scroll` callbacks.
+- Main-thread event-handler registration on `AppKitWindow`.
+
 ## [0.1.0] - 2026-04-18
 
 ### Added